Utilize Windows types

+++ ddebug/windebug.cpp
@@ -1054,7 +1054,7 @@ static void PrintFileTime(char *sztime, FILETIME ftime) {
 
 static void ShowModuleInfo(HANDLE LogFile, HINSTANCE ModuleHandle) {
   char FmtString[2000];
-  unsigned long bytesout;
+  DWORD bytesout;
   char ModName[MAX_PATH];
   __try {
     if (GetModuleFileName(ModuleHandle, ModName, sizeof(ModName)) > 0) {
@@ -1099,7 +1099,7 @@ static void ShowModuleInfo(HANDLE LogFile, HINSTANCE ModuleHandle) {
 
 static void RecordModuleList(HANDLE LogFile) {
   char FmtString[2000];
-  unsigned long bytesout;
+  DWORD bytesout;
   wsprintf(FmtString, "\r\n"
                       "Modules:\r\n");
   WriteFile(LogFile, FmtString, lstrlen(FmtString), &bytesout, 0);
@@ -1136,7 +1136,7 @@ static void RecordModuleList(HANDLE LogFile) {
 
 static void RecordSystemInformation(HANDLE LogFile) {
   char FmtString[2000];
-  unsigned long bytesout;
+  DWORD bytesout;
   FILETIME CurrentTime;
   GetSystemTimeAsFileTime(&CurrentTime);
   char TimeBuffer[100];
@@ -1218,16 +1218,16 @@ int __cdecl RecordExceptionInfo(PEXCEPTION_POINTERS data, const char *Message) {
     char *p = Debug_DumpInfo();
     lstrcpy(callstack, p);
 
-    unsigned long NumBytes;
+    DWORD NumBytes;
     SetFilePointer(LogFile, 0, 0, FILE_END);
 
     WriteFile(LogFile, topmsg, lstrlen(topmsg), &NumBytes, 0);
     WriteFile(LogFile, FORMATCRLF, lstrlen(FORMATCRLF), &NumBytes, 0);
 
     char Username[100];
-    unsigned long unamelen = 99;
+    DWORD unamelen = 99;
     char Machinename[200];
-    unsigned long cnamelen = 199;
+    DWORD cnamelen = 199;
     GetUserName(Username, &unamelen);
     GetComputerName(Machinename, &cnamelen);
     wsprintf(callstack, "Username: %s\r\nMachineName: %s\r\n", Username, Machinename);
